requirements.txt
pdfplumber
pydantic
python-dotenv
openai
langchain
langchain-openai
